+++ im_ui/src/main/java/com/test/yanxiu/im_ui/contacts/bean/ClassBean.java
@@ -9,11 +9,11 @@ import com.test.yanxiu.im_ui.contacts.DatabaseFramework.annotation.DbTable;
 @DbTable("yxClass")
 public class ClassBean {
     @DbPrimaryKey
-    private Integer classId;
+    private Long classId;
 
     private String className;
 
-    public ClassBean(Integer classId, String className) {
+    public ClassBean(Long classId, String className) {
         this.classId = classId;
         this.className = className;
     }
@@ -21,11 +21,11 @@ public class ClassBean {
     public ClassBean() {
     }
 
-    public Integer getClassId() {
+    public Long getClassId() {
         return classId;
     }
 
-    public void setClassId(Integer classId) {
+    public void setClassId(Long classId) {
         this.classId = classId;
     }
